move to olap.go
