+++ pkg/repository/sqlcv1/olap.sql.go
@@ -421,7 +421,9 @@ SELECT
     create_v1_olap_partition_with_date_and_status('v1_tasks_olap'::text, $2::date),
     create_v1_olap_partition_with_date_and_status('v1_runs_olap'::text, $2::date),
     create_v1_olap_partition_with_date_and_status('v1_dags_olap'::text, $2::date),
-    create_v1_range_partition('v1_payloads_olap'::text, $2::date)
+    create_v1_range_partition('v1_payloads_olap'::text, $2::date),
+    create_v1_range_partition('v1_otel_trace_olap'::text, $2::date),
+    create_v1_range_partition('v1_otel_trace_lookup_olap'::text, $2::date)
 `
 
 type CreateOLAPPartitionsParams struct {
@@ -1483,6 +1485,10 @@ WITH task_partitions AS (
     SELECT 'v1_cel_evaluation_failures_olap' AS parent_table, p::TEXT AS partition_name FROM get_v1_partitions_before_date('v1_cel_evaluation_failures_olap', $2::date) AS p
 ), payloads_partitions AS (
     SELECT 'v1_payloads_olap' AS parent_table, p::TEXT AS partition_name FROM get_v1_partitions_before_date('v1_payloads_olap', $2::date) AS p
+), otel_trace_partitions AS (
+    SELECT 'v1_otel_trace_olap' AS parent_table, p::TEXT AS partition_name FROM get_v1_partitions_before_date('v1_otel_trace_olap', $2::date) AS p
+), otel_trace_lookup_partitions AS (
+    SELECT 'v1_otel_trace_lookup_olap' AS parent_table, p::TEXT AS partition_name FROM get_v1_partitions_before_date('v1_otel_trace_lookup_olap', $2::date) AS p
 ), candidates AS (
     SELECT
         parent_table, partition_name
@@ -1544,6 +1550,20 @@ WITH task_partitions AS (
         parent_table, partition_name
     FROM
         payloads_partitions
+
+    UNION ALL
+
+    SELECT
+        parent_table, partition_name
+    FROM
+        otel_trace_partitions
+
+    UNION ALL
+
+    SELECT
+        parent_table, partition_name
+    FROM
+        otel_trace_lookup_partitions
 )
 
 SELECT parent_table, partition_name
