+++ pkg/repository/sqlcv1/tasks.sql
@@ -10,9 +10,7 @@ SELECT
     create_v1_range_partition('v1_event_to_run', @date::date),
     create_v1_range_partition('v1_durable_event_log_file', @date::date),
     create_v1_range_partition('v1_durable_event_log_entry', @date::date, 80),
-    create_v1_range_partition('v1_durable_event_log_branch_point', @date::date, 80),
-    create_v1_range_partition('v1_otel_trace_olap', @date::date),
-    create_v1_range_partition('v1_otel_trace_lookup_olap', @date::date)
+    create_v1_range_partition('v1_durable_event_log_branch_point', @date::date, 80)
 ;
 
 -- name: EnsureTablePartitionsExist :one
@@ -37,8 +35,6 @@ WITH tomorrow_date AS (
     SELECT 'v1_durable_event_log_entry_' || to_char((SELECT date FROM tomorrow_date), 'YYYYMMDD')
     UNION ALL
     SELECT 'v1_durable_event_log_branch_point_' || to_char((SELECT date FROM tomorrow_date), 'YYYYMMDD')
-    UNION ALL
-    SELECT 'v1_otel_trace_olap_' || to_char((SELECT date FROM tomorrow_date), 'YYYYMMDD')
 ), partition_check AS (
     SELECT
         COUNT(*) AS total_tables,
@@ -76,10 +72,6 @@ WITH task_partitions AS (
     SELECT 'v1_durable_event_log_entry' AS parent_table, p::text as partition_name FROM get_v1_partitions_before_date('v1_durable_event_log_entry', @date::date) AS p
 ), durable_event_log_branch_point_partitions AS (
     SELECT 'v1_durable_event_log_branch_point' AS parent_table, p::text as partition_name FROM get_v1_partitions_before_date('v1_durable_event_log_branch_point', @date::date) AS p
-), otel_trace_partitions AS (
-    SELECT 'v1_otel_trace_olap' AS parent_table, p::text as partition_name FROM get_v1_partitions_before_date('v1_otel_trace_olap', @date::date) AS p
-), otel_trace_lookup_table_partitions AS (
-    SELECT 'v1_otel_trace_lookup_olap' AS parent_table, p::text as partition_name FROM get_v1_partitions_before_date('v1_otel_trace_lookup_olap', @date::date) AS p
 )
 
 SELECT
@@ -156,20 +148,6 @@ SELECT
     *
 FROM
     durable_event_log_branch_point_partitions
-
-UNION ALL
-
-SELECT
-    *
-FROM
-    otel_trace_partitions
-
-UNION ALL
-
-SELECT
-    *
-FROM
-    otel_trace_lookup_table_partitions
 ;
 
 -- name: DefaultTaskActivityGauge :one
