+++ pkg/repository/sqlcv1/tasks.sql.go
@@ -236,9 +236,7 @@ SELECT
     create_v1_range_partition('v1_event_to_run', $1::date),
     create_v1_range_partition('v1_durable_event_log_file', $1::date),
     create_v1_range_partition('v1_durable_event_log_entry', $1::date, 80),
-    create_v1_range_partition('v1_durable_event_log_branch_point', $1::date, 80),
-    create_v1_range_partition('v1_otel_trace_olap', $1::date),
-    create_v1_range_partition('v1_otel_trace_lookup_olap', $1::date)
+    create_v1_range_partition('v1_durable_event_log_branch_point', $1::date, 80)
 `
 
 func (q *Queries) CreatePartitions(ctx context.Context, db DBTX, date pgtype.Date) error {
@@ -340,8 +338,6 @@ WITH tomorrow_date AS (
     SELECT 'v1_durable_event_log_entry_' || to_char((SELECT date FROM tomorrow_date), 'YYYYMMDD')
     UNION ALL
     SELECT 'v1_durable_event_log_branch_point_' || to_char((SELECT date FROM tomorrow_date), 'YYYYMMDD')
-    UNION ALL
-    SELECT 'v1_otel_trace_olap_' || to_char((SELECT date FROM tomorrow_date), 'YYYYMMDD')
 ), partition_check AS (
     SELECT
         COUNT(*) AS total_tables,
@@ -1338,10 +1334,6 @@ WITH task_partitions AS (
     SELECT 'v1_durable_event_log_entry' AS parent_table, p::text as partition_name FROM get_v1_partitions_before_date('v1_durable_event_log_entry', $1::date) AS p
 ), durable_event_log_branch_point_partitions AS (
     SELECT 'v1_durable_event_log_branch_point' AS parent_table, p::text as partition_name FROM get_v1_partitions_before_date('v1_durable_event_log_branch_point', $1::date) AS p
-), otel_trace_partitions AS (
-    SELECT 'v1_otel_trace_olap' AS parent_table, p::text as partition_name FROM get_v1_partitions_before_date('v1_otel_trace_olap', $1::date) AS p
-), otel_trace_lookup_table_partitions AS (
-    SELECT 'v1_otel_trace_lookup_olap' AS parent_table, p::text as partition_name FROM get_v1_partitions_before_date('v1_otel_trace_lookup_olap', $1::date) AS p
 )
 
 SELECT
@@ -1418,20 +1410,6 @@ SELECT
     parent_table, partition_name
 FROM
     durable_event_log_branch_point_partitions
-
-UNION ALL
-
-SELECT
-    parent_table, partition_name
-FROM
-    otel_trace_partitions
-
-UNION ALL
-
-SELECT
-    parent_table, partition_name
-FROM
-    otel_trace_lookup_table_partitions
 `
 
 type ListPartitionsBeforeDateRow struct {
